requirements.txt
Django>=4.2.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
google-cloud-translate>=3.0.0
gunicorn>=21.2.0
whitenoise>=6.5.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
psycopg2-binary>=2.9.9  # For PostgreSQL support
python-magic>=0.4.27    # For better file type detection
Pillow>=11.0.0          # For image processing
django-storages>=1.14.2 # For cloud storage (S3, etc.)
boto3>=1.34.0           # For AWS S3 integration
sentry-sdk>=1.40.0      # For error tracking
django-cors-headers>=4.3.1 # For handling CORS
django-redis>=5.4.0     # For caching
dj-database-url>=2.1.0  # For database URL configuration
django-filter>=23.5     # For filtering querysets
django-cleanup>=8.0.0   # For automatic file cleanup
django-compressor>=4.4  # For compressing static files 